moved max_rows_per_split computation to split module
